Cropping logic added
